Wide and Deep Network Architectures
Prediction task to perform on heart dataset, evaluate two different deep learning architectures and tune hyper-parameters for each architecture. It is helpful to have a large dataset for building a wide and deep network. It is also helpful to have many different categorical features to create the embeddings and cross-product embeddings.