Classification Task & Exploratory Data Analysis [EDA]
This notebook book loads csv file, perform exploratory analysis to understand meaningful patterns. It modifies target to use this dataset for classification task. It also perform preprocessing task required to be performed before feeding data to the model. Performance visual for the model is provided to show relevance of the model to the dataset. This Jupyter Notebook is completed using Python and its ML libraries namely; Pandas, NumPy, matplotlib, plotly, Seaborn, Scikit-learn.